requirements.txt
# FastAPI核心
fastapi>=0.104.1,<0.120.0
uvicorn[standard]>=0.24.0,<0.35.0

# 数据库相关
sqlalchemy>=2.0.23,<2.1.0
pymysql>=1.1.0,<1.2.0
aiomysql>=0.2.0,<0.3.0
cryptography>=41.0.7,<43.0.0

# Pydantic和数据验证
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# 密码加密
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt>=4.0.0,<5.0.0

# HTTP客户端
httpx>=0.25.2,<0.28.0

# Nacos服务注册与发现
nacos-sdk-python>=0.1.5,<0.2.0

# 工具库
python-multipart>=0.0.6,<0.1.0
python-dotenv>=1.0.0,<2.0.0
apscheduler>=3.10.4,<4.0.0

# JSON处理
orjson>=3.9.0,<4.0.0

# 日志
loguru>=0.7.0,<0.8.0

# 异步支持
asyncio-mqtt>=0.16.0,<0.17.0

# 开发和测试工具
pytest>=7.4.3,<8.0.0
pytest-asyncio>=0.21.1,<0.24.0
pytest-cov>=4.1.0,<5.0.0
faker>=20.1.0,<25.0.0

# 代码质量
black>=23.0.0,<25.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<8.0.0